debug

+++ Modules/MachineLearning/3D_resnet.py
@@ -118,7 +118,7 @@ if args.mode == 'train':
 
 		for batch_idx, (data, target, path) in enumerate(trainset_loader):
 			target = target.cuda(async=True)
-			data = data.float()
+			# data = data.float()
 			data = Variable(data)
 			target = Variable(target)
